refactor(xtask): split bundle.rs along what it actually does
517 lines held five separate jobs: rendering the icon, embedding the helpers
and CLI, stamping Info.plist keys, signing, and the assembly order that drives
all of them. Only the last one is what `bundle` means, so the other four move
into modules named after themselves and `bundle.rs` is down to 126 lines.

The rest of the macOS domain still reaches `bundle::HELPERS`,
`bundle::generate_icns` and `bundle::quoted_identity`, so dev_bundle, dmg and
macos.rs need no edit: the facade re-exports them. `stamp_bundle_version` did
not survive that treatment — nothing outside `bundle` ever called it, so its
`pub(super)` was a dead export and it is now scoped to where it is used.

Tests move next to what they pin: the helper-table and required-binary tests to
`embed/tests.rs`, the entitlements test to `signing/tests.rs`. Count is
unchanged at 46.

rustdoc caught the one thing the compiler could not: a `[`identity::stamp`]`
link that resolved from bundle.rs and does not from embed.rs.
